backend/api/requirements.txt
fastapi
uvicorn
pyjwt
passlib
bcrypt
sqlalchemy[asyncio]
python-dotenv
python-jose
python-multipart